DOSE-121 ingest to zettacache on write (#412)
Ingest written blocks to the zettacache.  Note that this is an on-disk format
change due to some code cleanup.  Several synergistic improvements as well:

don't allow overwrites for sync-to-convergence; clean up now-unused code in
Agent

limit disk i/o queue depth.  This prevents starvation of the 512 Tokio
blocking threads.  Otherwise we could use up all of these threads with the
write_raw() blocking task that does the pwrite() syscall.

separate disk i/o queue depth limits for reads vs writes.  This prevents
starvation of reads (for cache hits) when the writes (for inserts) are at the
max depth

limit buffering of cache insertions.  When the buffer is full, the insertion
is dropped (ignored).  When the cache is slower than the network (which it is
when ingesting writes, but not when inserting from cache misses), we don't
want to wait for the cache, slowing down overall write performance; instead
we buffer a small amount (about a second's worth) and once that limit is
reached we don't insert the next block to the cache.  Victim blocks are
selected randomly (based on arrival order).  In the future we might want to
have the kernel tell us which writes contain data that's more likely to be
read in the future (e.g. ZFS metadata).

block allocator: continue looking from last allocation location.  This vastly
reduces CPU usage, once evictions have happened and the free space is
fragmented.

TARGET_CACHE_SIZE_PCT default to 80

more destructuring code cleanup in range_tree.rs

zettacache: remove unnecessary chunk_summary

fix DOSE-602 cross-PendingFreesLog object consolidation breaks pool