type checking for {Object,Block}BasedLogPhys (#432)
We can have additional type safety by using generics and PhantomData to associate the type of the ObjectBasedLogPhys and BlockBasedLogPhys with the type of the entries stored in it. This prevents us from loading an ObjectBasedLogPhys that stores one type of entries into a ObjectBasedLog that stores a different type of entry (same for BlockBasedLog).