DOSE-469 agent crash can leak deleted objects (#449)
For reclaiming free space, as it completes, we need to write the txg
that records our progress, and then after that delete the DataObject's
that are no longer needed ("obsolete objects"). If we crash between
those 2 operations, we lose track of the need to delete the objects and
they will be leaked.

This PR adds a new field to the UberblockPhys that lists the "obsolete"
DataObject's (ObjectId's) that need to be deleted after the txg
completes. If the agent crashes, we redo the deletion.