track and check last_entry (#445)
When inserting a newly-written block into the zettacache, surprisingly,
the BlockId may be in the cache, if the system crashed or the pool
rewound, causing a BlockId that was already persisted to the cache to be
reused.

This PR forces all cache insertions to check in the cache first. The
performance of this is improved by storing the "last entry", so we don't
have to read from the on-disk index if it's after that. Note that this
only helps when there's only one pool's data in the cache (which will
normally be the case for the product but may not be during testing).