delete and list objects using Stream (#472)
Change the way that the ObjectAccess layer lists and deletes objects to
use Streams (which are like async iterators).  This allows entries to be
processed as they become available, and allows for more elegantly
manipuating the list/stream of object names (keys).  Note that this
doesn't introduce additional concurrency (except as noted below), but it
reduces copying around of the lists of keys (i.e. there's no
Vec<String>, just a series of String's that are moved from producer to
consumer).

Consolidate the consumers of the various listing helper functions around
just 2 calls: list_objects() and list_prefixes().

The only delimiter used is "/"; change the parameter to be a boolean
indicating whether it should be used.

Note: the concurrency in `cleanup_data_objects()` (listing all prefixes
at the same time) is maintained by use of `select_all()`.  Additionally,
we now do a separate, concurrent DeleteObjects call for the contents of
each prefix.

Side benefit: "list pools" command is 10x faster, by virtue of reading
all the pools uberblocks concurrently, using
`StreamExt::for_each_concurrent()`.  This required introducing a lock
around the results, since each pool may be added to the results
concurrently.  (However, the time for a scanning `zpool import` is often
dominated by serially opening each valid pool.)

Note: this probably introduces even more bugs with dealing with a
non-empty AWS_PREFIX.  A future PR will remove this functionality.