die at specific location (#489)
Allow specifying a specific file and/or line to die at, with the maybe_die_with() infrastructure.

Caveat: maybe_die_with() must be called at least once from the selected call site before the time that we choose to die, otherwise the infrastructure won't be aware of that potential call site, and we won't die at all. We don't attempt to re-check later.