fix zettacache locking (#484)
The zettacache buffers insertion requests in memory, up to a certain
limit, writing them to disk as the disk can keep up.  Once an insertion
is assigned to a specific checkpoint (which is always "the current
checkpoint" - there's no open/quiescing/syncing TXG train), that
checkpoint needs to wait for the insertion to complete (including its
write to disk) before the checkpoint can be completed.  This ensures
that if we crash, the logical index (typically the operation log)
doesn't refer to disk locations that weren't actually written.

The problem is that we assign an insertion to the checkpoint early -
when it's buffered.  (This happens when we add it to
ZettaCache::outstanding_inserts.)  It may remain buffered for a long
time (dozens of seconds), depending on CACHE_INSERT_MAX_BUFFER and
underlying disk performance.

When writing out a checkpoint, we have the ZettaCacheState lock, and we
wait for the insertions assigned to this checkpoint to complete (the
outstanding_inserts).  If this takes a long time, it can be very bad for
performance because new lookups and insertions have to wait (many
seconds) for the zettacache state lock.

The solution implemented in this PR is to wait until later to assign
insertions to a checkpoint, waiting until we are able to issue a write
to disk.

As part of a separate PR, we may want to reduce the impact of the (now
small, ~1 second) time it takes to sync a checkpoint.  For example by
having lookup/insert requests fail rather than wait for the state lock.
Or to have a open/syncing train like we do for TXG's, where we can
accumulate more changes in "open context" while we're in the middle of
writing out a batch of changes from "syncing context".