Waiting for DeleteObjects can postpone txg (#1)
In Pool::end_txg(), we wait for the deletion of objects that were
obsoleted in the previous txg.  This can become a performance problem,
since it may take dozens of seconds, but we need to push a txg every few
seconds to keep up with the write workload.  This system can be nearly
idle for dozens of seconds while waiting for these object deletions to
complete.

This PR changes the behavior such that the object deletion can take
place over many TXG's. We can add more objects to the "delete queue"
(list of obsolete objects) while still processing the previous batch.
And we can record the deletion progress each txg so that if we crash, we
don't have to re-issue many completed deletions.

Side benefit: refactored the object deletion code into its own file, to
slightly reduce the size of pool.rs.