switch to 2021 edition of Rust (#21)
Make 2 changes that are not strictly required but make sense given the
changes in the 2021 edition:

1. TryFrom/TryInto no longer need to be imported, as they are part of
the standard prelude.  Remove the redundant `use` statements.

2. When possible, closures capture members of local variables, rather
than the entire variable.  This can cause the non-captured members to be
dropped later than in the 2018 edition.  In our case (in
ZettaCache::open()), this doesn't really matter, but it brought to my
attention that we're cloning and then discarding a bunch of the members
of ZettaCache (without using them).  So I changed this to just clone the
few members that are actually needed by the closures.