Fix calculations for target cache size in try_start_merge_task() (#35)
The calculations done by try_start_merge_task() to determine if we
should start a merge, and potentially trigger evictions, currently uses
the block access object to determine the total size of the cache.

This is incorrect, as that size includes the amount of space reserved
for various forms of metadata used by the cache. What we actually want
to use, is the total size available to the block allocator object, as
that correctly excludes all the space reserved for cache metadata.