halve ObjectBlockMap RAM by using ObjectId = min_block (#54)
The memory footprint of the ObjectBlockMap can be cut in half by
treating the ObjectId as the minimum BlockId that it could contain. Now
the ObjectBlockMap only needs to be a sorted list of ObjectId's, rather
than (ObjectId, BlockId) tuples.

In other words, rather than having the ObjectId's be 0, 1, 2, 3, ...
they are 0, (first BlockId in the second object), (first blockId in the
third object), etc. Since we already treat ObjectId's as potentially
non-sequential (due to object consolidation), the implications on the
rest of the code are minimal.

Rather than rotating to a new ReclaimLog every 1000 ObjectId's, we now
rotate every 1,000,000 ObjectId's (i.e. every 1,000,000 BlockId's).

side cleanup: ObjectDeletePhys -> ObjectDeleterPhys