Streams cleanup (#56)
With regards to Streams, and specifically Streams of Futures, the
relationship of .buffered() and Futures[Un]Ordered wasn't clear to me
when I started using them. There's substantial simplification that we
can do in this area.

.buffered() consumes a Stream of Futures. This stream can be created
more simply from an iterator of Futures; FuturesOrdered need not be
involved.

In some cases, the stream! macro isn't needed, we can implement the
stream by using the stream combinators in StreamExt.