zettacache whole-object ingestion (#115)
When we are writing new blocks, or doing sibling block ingestion, we try
to insert all the blocks of an object to the zettacache. When the block
device can't keep up with the ingestion rate, the buffer will be full
and this results in some blocks from each object being ingested. This is
not great because if we read the file back later, we'll need to read
every object from S3, despite having some of its blocks in zettacache.

This commit makes us ingest all the blocks in an object, or none of
them, when trying to ingest all the blocks of an object. This allows
some additional optimizations along the way.

When converting Bytes to AlignedBytes, check if it's already aligned.
This is helpful when doing ingestion to zettacache from a DataObject
whose blocks came from "write block" commands, which have already been
aligned.