index chunk cache (#114)
Add a cache of chunks of the index, at the BlockBasedLogWithSummary
layer. The big win here is being able to determine that a block is NOT
in the cache without reading from disk. This is needed primarily when
ingesting newly-written blocks, and when doing sibling block ingestion.
In both of these cases, we have to check if the blocks are already
present, but we are checking many adjacent BlockId's, whose entries
would fall in the same chunk of the index. The cache also eliminates the
need for the last_entry special case.