lookup_impl() must hold locked key (#129)
When doing a lookup, we should ensure that a LockedKey 
is held, by requiring that it be passed into ZettaCache::lookup_impl() 
and ZettaCacheState::lookup(). If the key wasn't held, a 
lookup could change the entry's atime in the middle of an 
insert which could cause the atime histogram to be out of 
sync with the pending changes.