fix resuming of in-progress merge (#128)
If a checkpoint is created while a merge is in progress but before
getting the first MergeMessage, the checkpoint code doesn't know the
ZettaCacheIndexPhys of the new index, so it creates a new, empty index.
However, this empty index has the wrong atime_histogram.first_ghost and
first_live (they are zero). Typically, this is not noticed, because in
the next checkpoint we'll get the ZettaCacheIndexPhys and replace the
incorrect one. However, if the agent restarts before that happens, then
when we resume, the wrong first_ghost/first_live will cause the resumed
merge to proceed incorrectly.

This commit fixes this by passing around the new ZettaCacheIndexPhys
along with the Receiver used to communicate with the merge task.