improve performance of lock_non_send (#125)
Non-send locks are implemented with an async_trait, which is ergnomic to
use, but unfortunately requires an additional memory allocation for
every call to the async trait.  This was observed to have a performance
impact (from inspecting CPU flame graphs under a moderate workload).

This change removes the use of async_trait, instead using a plain
function to do the non-send lock.