DOSE-872 invalid zettacache state due to losing UpdateAtime during merge (#122)
This change eliminates the explicit remove capability (PendingChange::Remove). The heal() function now just insert()s a new value for the same key. In the merge code a request to insert when an entry already exists is treated as a replacement request (and the older entry is evicted). This simplifies a number of code paths. With the removal of PendingChange::Remove, PendingChange::RemoveThenInsert is also unnecessary. Note also that the operation log now only consists of Insert entries.