measure memory usage of our data structures (#108)
This changes the memory allocator to wrap the default one and track
allocations based on a tag (string literal) set in a thread-local
variable. When an allocation is performed, we increment a counter based
on the tag, and append the tag to the allocation. The allocation is made
8 bytes larger to store the tag. When a free is performed, if the tag is
present then we decrement a counter based on the tag.

We log the memory used by each tag once a minute.

The performance overhead with the default settings is minimal, because
only low-frequency allocations are tagged. To investigate memory usage
by untagged allocations, backtrace-based tagging can be enabled, which
is much more expensive even when only sampling one allocation per ~8KB
allocated.

Future work could include accumulating the counters in per-thread
structures, to avoid the global lock on the counters, which might let us
enable tagging of high-frequency allocations in production.