Atime should be u32 (#153)
The Atime should be u32 rather than u64.  In combination with
`repr(packed)`, this reduces the size of a PendingChange from 32 to 24
bytes.

This is not an on-disk change, because the Atime is represented on disk
as either a decimal number (JSON) or variable-length binary integer
(bincode).

Given the current 10-seconds per Atime value, a cache can live for
1361 years before the Atime would wrap.  Our original design called for
a 32-bit atime.