DOSE-1000 Update user/kernel API to improve error handling
The user/kernel API uses nvlists to communicate.  Each request and
response is free to construct the nvlist as they please.  This results
in varied methods of passing parameters, in particular error values.

This commit changes the API to standardize the error handling.  Note
that only error responses are modified.  Other calls can later opt in to
returning errors without changing the API of successful responses.