Reduce size of UpdateAtime by packing members. (#151)
Pack members of DiskLocation and IndexValue, allowing them to be
unaligned.  Make the members private so that we can be sure that the
compiler knows about all potentially unaligned memory access.  This will
also facilitate future changes to the in-memory layout of
PendingChanges.

size_of<PendingChanges> is reduced to 57% of its former size (56 -> 32 bytes)