DOSE-987 Use new index during merge (#145)
If the agent restarts during a merge, the PendingChange::UpdateAtime
records that have been persisted to the new index are not present in the
pending_changes after restarting.  This causes several inconsistencies
when the merge completes and switches from the old index to the new
index.

This commit addresses the problems by using the part of the new index
that's already been written, while the merge is still in progress.  This
way, if the agent restarts, we lose only UpdateAtime records that are
not represented anywhere on disk.

To accomplish this, ZettaCache::lookup() needs to look in either the old
or the new index, depending on if this key has already been merged to
the new index.  A readonly view of the new (in progress) index is made
available to the lookup() code.  A new readonly view of a
SummarizedBlockBasedLog is introduced to support this.  The readonly
view is periodically updated as the merge progresses, to reflect the
newly appended chunks/entries.

Additionally, several structs/members are renamed to be more consistent:

BlockBasedLogWithSummary -> SummarizedBlockBasedLog
ZettaCacheIndex -> IndexRun
index -> new_index or old_index
AtimeHistogram::first() -> first_ghost()

Use nice_p2size() throughout zettacache.rs.

Make IndexRun fields non-pub, introduce accessor methods.

Introduce type for PendingChanges (BTreeMap).