Use a separate thread to read from BlockBasedLog. (#152)
When reading from a BlockBasedLog (e.g. when merging and reading the old
index), we want the reading thread (e.g. merge_task()) to be able to get
entries without delay, using the other resources of the system (disk,
CPU) in the background so that the reading thread can go as quickly as
possible.  To accomplish this, we introduce 2 new tasks, which can run
on separate threads.  This way we offload the CPU cost of deserializing
the BlockBasedLogChunk's to a separate thread, and we can be
concurrently reading from disk, deserializing, and the merge_task can be
consuming the entries.

Merging speed is increased from 4.69 million entries/sec to 6.74 million
entries/sec, in increase of 44%.