New layout of DataObjectPhys. (#168)
All old pools will need to be destroyed, as the old on-disk format is no
longer supported.

The new DataObject has 3 sections:
1. size word: encoded length of DataObjectPhys (plus pad bits for future
   use, e.g. verioning)
2. bincode-encoded DataObjectPhys, which contains BlockId's and block
   sizes, not block contents.  Entries are sorted by BlockId.
3. block contents, in order specified by DataObjectPhys.  By keeping
   this separate from the DataObjectPhys, we can do a sub-object
   Get of just the first 2 sections and decode them on their own.

The DataObjectPhys has the BlockId's and block sizes encoded as byte
arrays, so that they can be decoded by serde in constant time.  Code is
added to access/byteswap each entry as needed.

A new put_object_stream() method is added which takes a ByteStream,
allowing us to put a DataObject to S3 without copying each of its blocks
into a contiguous buffer.

When reading a single block, a fast path is added to avoid constructing
the entire DataObject; only O(log(n)) entries need to be
accessed/byteswapped to binary search for the target block.  In the
future this can be further enhanced to do a sub-object Get of just the
DataObjectPhys (or find the offset in the ZettaCache), followed by a
sub-object Get of just the target block.

bonus changes:
* rename ingest_all() -> insert_all() to match insert()
* change some trace! to super_trace!