limit number of outstanding Get/Put operations by type (#182)
Limit the number of Get/Put operations that the agent does concurrently.

The limit is applied to each ObjectAccessStatType separately.  This
ensures that even if there are a lot of concurrent operations on
DataObject's, we will still send out operations on metadata, which may
be particularly important for the heartbeat.

The limit also increases overall performance, by reducing CPU contention
due to locking and scheduling.

Microbenchmarks (outside of ZFS) show that with 2MB objects, a queue
depth of 25 should be sufficient to saturate a 25Gbps network.  With
ZFS, performance continues increasing up to a queue depth of around 50,
with throughput around 17Gbps.  The default queue depth is set a bit
higher than this to provide some headroom on larger instances.

A workload that exercises this well (i.e. can issue many object Get's
concurrently) is sequentially reading (with prefetching) a large file
that was randomly written.  Or randomly reading a large file, from many
threads concurrently.