DOES-957 Iterating over the index cache while remapping is slow (#139)
This commit changes the model for index_cache updates. Instead of waiting
until the end of the merge, the key's associated with remapped data are
collected during the merge and passed to the checkpoint_task periodically
so that the index_cache can be updated. This changes an
O(size of index_cache) process into an O(number of remapped blocks)
process. In addition, the cost of that process us spread across the entire
merge cycle.