DOSE-1002 Add a pool GUID table to ZettaCacheState (#177)
Currently, all index keys stored in the zettacache contain a pool GUID. 
A GUID is a 64bit value. We can reduce this to an 8bit value by using 
a mapping table (limited to 256 entries). We can store the mapping table 
in the checkpoint and just use an 8 bit table index as a PoolId in the 
IndexKey.

Note: this changes the representation of an index key on disk, 
so will invalidate old caches.