Use Derivative instead of TerseVec
Use `#[derivative(Debug(format_with = "util::tersevec"))]` instead of
the TerseVec type to format large Vec's for Debug.  This encapsulates
the special case within the struct definition.  Consumers of the struct
don't need to care about the Vec being formatted tersely or not.

The only functional change here is to tersely format the
ObjectDeleterPhys, which was seen to have over 1 million entries on a
test system.