read & write threads (#194)
The low-level read and write calls to the disk are done with blocking
pread() and pwrite() system calls, so we can't do them from regular
async Tokio tasks, as that would block the worker thread.  Instead we
dispatch these blocking calls to Tokio's blocking threads.

The problem is that the performance of creating blocking tasks with high
frequency is not great, due to CPU overhead of the tokio scheduler, and
also the overhead of putting threads to sleep and waking them up in
userland (using the futex() syscall).

For example, consider the case where there are always i/o's to process,
and we are at the maximum queue depth (QD) (i.e. there are no permits
left in Disk::outstanding_reads/writes, and tasks are awaiting permits).
There will be QD blocking worker threads each processing one of the
blocking tasks to do an i/o.  When a blocking task completes, there is
no additional task for it to work on (because the limit of the number of
blocking threads is greater than the number of concurrent blocking
tasks), so it has to go to sleep.  The initiating task will get the
response and drop the semaphore, allowing another task to dispatch
another blocking task, waking a blocking worker thread.

We can improve performance a bit by avoiding this frequent sleep-wake of
the worker threads, by using our own dedicated threads for pread() and
pwrite(), sending them messages to indicate what syscalls we need to
make.  When a worker completes pread/pwrite, it may find that there is
more work to do (another message in the queue) so it can start on that
immediately without going to sleep.

We use a Flume MPMC (multiple producer, multiple consumer) channel for
the messages.  The number of threads is the same as the configured
maximum queue depth.

Performance results (on i3en.2x, local NVMe SSD):
  Sequential read of 2 cached files:
  55K -> 75K IOPS (+36%)
  Random reads of uncached file with sibling block ingestion:
  60K -> 68K IOPS (+13%)

Bonus cleanup:

* cleaner encapsulation of `unsafe` functionality in block_access.rs
* more error handling in Disk::new()
* lowest level of reading/writing moved into `impl Disk`