Fix error with 3+ zettacache devices. (#218)
There was a bug introduced by #194 where if there are 3 or more
zettacache devices, we can't communicate with S3.  The problem is that
we've created many threads with tokio::task::spawn_blocking() (192 per
disk), to handle read/write i/o's to disk, and we've run into the
default limit of 512 tokio blocking threads.  Then when rusoto/reqwest
wants to create a blocking thread, it hangs and eventually hits a
timeout.

The fix is to use std::thread::spawn(), which doesn't have a limit
(beyond what's imposed by the OS, 58344 on a small VM).