Increase index chunk size. (#219)
Increase the index chunk size to 8KB (337 entries per chunk).  This will
decrease the memory usage of the chunk summary, as well as the time it
takes to read it from disk when the agent starts.

The chunk is required to be a multiple of the sector size (512B
currently, but will be 4KB on other platforms).  The index chunk now
uses a multiple of 4KB on disk, with a minimum of padding at the end of
the chunk.

The JSON-encoded header on the beginning of each block is minimized by
using one-letter field names and enum variant names.  The old names are
accepted for backwards compatibility.