DOSE-1039 loading spacemaps is prohibitively slow (#215)
When the agent starts, it first loads the block allocator data from the
spacemaps.  If the spacemaps are large, this can take a long time.  The
spacemaps can be larger than necessary if they contain many allocs/frees
that are adjacent or overlapping.  The overlapping case is controlled by
the SLAB_CONDENSE_RATE_FACTOR tunable.  However, even if there are few
overlapping allocs/frees, there can be a huge number of spacemap entries
(on the order of as many blocks are in the cache, ~1 billion per 3TB).

This commit introduces a new factor for determining the condensing rate,
SLAB_CONDENSE_MAX_BADNESS_RATIO, which takes into account adjacent
allocs/frees that could be represented with a single entry in the
spacemap.  The end result is that this increases the rate at which we
condense slabs, which reduces the size of the spacemaps, which reduces
the time it takes to load them.  The increased condensing activity is
not very expensive compared to other things we do each checkpoint with
the zettacache state lock held (we can condense >1000 slabs/ms).