DOSE-1045 failed to get object due to object/block map changing (#224)
The object-block mapping can change while a GET is in progress.  After
the object-block map changes, the old object can be deleted.  If the GET
is still in progress, the GET may fail with "object does not exist", and
the agent will panic.

The fix is to retry failed GETs if the object-block mapping for the
requested block has changed (i.e. if we will be getting a different
object).