DLPX-79806 [Backport of DLPX-79805 to 6.0.13.0] Checksum errors in object-store pool and healing activity in zettaCache (#238)
We sometimes see zettacache healing, and zpool checksum errors, even
when there has been no kernel panic that could trigger DOSE-614.  The
problem is that in #193, we removed the interlock between the disk
writes that happen as part of cache insertion, and disk reads that
happen as part of cache lookup.  So we can read a block before its write
has completed.  With sibling ingestion (which is off by default), it
could be relatively easy to get a lookup on a recently-inserted block.

This commit reintroduces an interlock by extending the hold time of the
LockedKey until the write to disk completes.  The performance of this is
better than the old implementation because we have to create the
LockedKey anyway.