ZettaCache::insert(): only memcpy() if actually inserting (#252)
In the common case calling ZettaCache::insert() (with the default of
SIBLING_BLOCKS_INGEST_TO_ZETTACACHE=false), we memcpy() the block to
avoid pinning the whole object in memory.  However, when doing
sequential reads, we may be dropping most insert requests due to the
insertion buffer being full.  In this case, the memcpy() and associated
buffer allocation is unnecessary work as it will be thrown away after we
realize the insertion will fail.  This has been observed to take >17% of
one CPU.

This commit changes ZettaCache::insert() to take a closure which
computes the AlignedBytes to insert, and we only invoke it if we
successfully get an insertion permit (because there's enough space in
the insert buffer).