Stop using timeout(1) for perf data collection
The utilities used to gather performance data are launched to run
indefinitely. timeout is used to stop them when the IO load is
finished. This can cause data to be lost if buffers aren't flushed to
stdout when the termination signal is received. We can avoid this issue
by launching the utilities directly for bounded time and dropping the
use of timeout.