use lazy_static_ptr! to store pointer to lazy_static globals for debugging (#277)
We would like to be able to print the log messages from a core file.
Unfortunately, lazy_static! makes it hard to get to the global object
from the debugger.  We would like to save the pointer in a global
variable so that we can easily dereference it from rust-gdb.

The way we are saving LOG_MESSAGES_PTR currently is not right; we're
saving the location of the variable on the stack, which will be moved
into the "real" location after it is returned.

This commit solves the problem by allocating the data on the heap
(inside a `Box`), so that we can know its address before the initializer
returns.  Because this is a little tricky, and you have to name the
(potentially complicated) type of the variable 3 times, a new macro is
introduced to do this for us.

Now in the debugger you can do:

```
(gdb) p util::logging::LOG_MESSAGES_PTR.p.value.data
$1 = core::cell::UnsafeCell<alloc::collections::vec_deque::VecDeque<alloc::string::String, alloc::alloc::Global>> {value: VecDeque(size=28) = {
...
    "[2022-03-04 23:01:49.089][zfs_object_agent][ERROR] Starting ZFS Object Agent (zfs-0.7.0-6003-g5b402b78c).  Local timezone is +00:00 (+00:00)",
```