tunable! macro (#267)
Introduce a new macro `tunable!` which is similar to `lazy_static!` (which it
uses internally), but is used for declaring configuration parameters.  The
primary advantage is not having to match the variable name
(SCREAMING_SNAKE_CASE) with the string that's used in the config file
(snake_case).  The syntax is the same as `lazy_static!`, but at runtime the
variable's value may be overridden by by the config file that was loaded with
`read_config()`.

Additionally, a `Tunable` trait is added for performing arbitrary conversion
from the values in the config file (which are parsed with the `Deserialize`
trait) to the types of the tunables.

All tunables are changed to use the new mechanism, which includes specialty
conversions for:

* `Duration`: tunable names use the `_duration` suffix, and are parsed from
human readable string values, e.g. `checkpoint_interval = 12345ms`.
* `ByteSize`: tunable names use the `_size` suffix, and are parsed from either
numbers or human readable strings, e.g. `write_aggregation_size = "128 KiB"`.
Note that `KB` is 1000 and `KiB` is 1024.
* `Percent`: tunable names use the `_pct` suffix, and are parsed from floating
point numbers. e.g. `target_cache_size_pct = 96.5`.

Unit tests are added for the tunable infratructure.